fix: ensure node crypto polyfill is applied early
